Added execOnMainThread to allow consumers to toggle between executeSqlBatch and backgroundExecuteSqlBatch
This is useful for apps which run in the background - for example, and app which is awakened to handle bluetooth events and persist them to SQLite.

Background activity throttling on iOS has been observed to suspend the app as soon as the main ReactNative thread is inactive, resulting in a hanging SQLite query which only completes several minutes later when the app next wakes.

It is now possible to monitor AppState and switch the exec method when the app is backgrounded.

While backgrounded the concerns about slow UI performance caused by locking the main React thread are not an issue.